[feat] bedrock mapper
